// projects/website/src/app/pages/product-detail/models/getProduct.response.ts
// Generated by https://quicktype.io

export interface GetProduct {
  message: string;
  data: Datum[];
  statusCode: number;
  status: boolean;
}

export interface Datum {
  _id: string;
  name: string;
  parentId: null;
  shortDesc: string;
  fullDesc: string;
  images: string[];
  status: boolean;
  sizes: string[];
  rating: number;
  SKU: string;
  mrp: number;
  quantity: number;
  sellingPrice: number;
  deleted: boolean;
  isCustomize: boolean;
  brand: BrandClass;
  category: CategoryClass;
  subCategory: CategoryClass;
  color: string;
  createdAt: string;
  updatedAt: string;
  slug: string;
  __v: number;
  variants?: string[];
}

export interface BrandClass {
  _id: string;
  brandName: string;
  isApproved: boolean;
  deleted: boolean;
  image: Image;
}

export interface Image {
  favicon: string;
  icon: string;
  lg: string;
  md: string;
  original: string;
  sm: string;
  x2lg: string;
  xlg: string;
  xs: string;
}

export interface CategoryClass {
  _id: string;
  status: boolean;
  deleted: boolean;
  name: string;
  image: string;
}